requirements.txt
# Core Python scientific stack
numpy
pandas
matplotlib
scipy

# Control systems
control          # Python Control Systems Library
sympy            # Symbolic math (Laplace, equations)

# Machine Learning
scikit-learn
torch            # or tensorflow (pick one ML framework)
torchvision      # if using PyTorch
numpy
pandas
matplotlib

# Reinforcement Learning (optional for ML controllers)
gym
stable-baselines3

# File handling & parsing
openpyxl          # Excel file handling (read/write .xlsx)
python-docx       # Word document handling (.docx)
PyPDF2            # Basic PDF reading and merging
reportlab         # PDF creation and text rendering
Pillow            # Image processing (load, edit, save images)
PyYAML            # YAML file handling (read/write .yaml)

# Web scraping & automation
requests
beautifulsoup4
selenium

# Utilities
tqdm             # Progress bars
pytest           # Testing framework

# Database
mysql-connector-python

